fix: Correct 'sign compare' warning

+++ Extensions/RosettaTorch/Sources/RosettaTorch/TorchPolicy.cpp
@@ -67,7 +67,7 @@ TaskMeta TorchPolicy::RequireMulligan(Player& player)
 
     // Generate mulligan action.
     HandZone& hand = player.GetHandZone();
-    for (size_t i = 0; i < hand.GetCount(); ++i)
+    for (int i = 0; i < hand.GetCount(); ++i)
     {
         actions.emplace_back(TaskID::MULLIGAN, hand[i]);
     }
